v.1.1.4: improved/fixed CVCT_MASTER handling
CVCT_MASTER can target arguments in addition to the token name, and it does so by checking the first arguments. So [BODY:HUMANOID_SIMPLE:3_EYES] would be targeted by CVCT_MASTER:HUMANOID_SIMPLE, but not by CVCT_MASTER:3_EYES. However, it does check for whole arguments, not for a string, so CVCT_MASTER:HUMANOID would not target the former.